refactor: split queryParameters.js into three modules
Discovery, ranges and values were three concerns in one 861-line file. Tests split to match. Pure code motion: every declaration and every test body is byte-identical.
