+++ src/js/utils/queryParameters.js
@@ -497,10 +497,11 @@ function inlineDataSlots(tree, text, prologue, wanted, into) {
   });
 }
 
-// ── the parameters, and their ranges ──────────────────────────────
+// ── the parameters a query offers ─────────────────────────────────
 //
-// What the form is built from: one entry per declaration, and the two
-// ends of a range where the query or its author says there is one.
+// What the form is built from: one entry per declaration, with the
+// values it may be given in place of. Which of them are the two ends of
+// a range is asked in parameterRanges.js.
 
 /**
  * The values a query offers to have changed.
@@ -543,319 +544,4 @@ export function queryParameters(tree, doc) {
     if (slots.length > 0) parameters.push({ variables, label, slots, declaredRange });
   }
   return parameters;
-}
-
-/**
- * How to read a pair of comparisons as a sentence.
- *
- * Each end of a range is inclusive or not, and English has words that say
- * so: "between the 1st and the 31st" takes both days in, "after the 1st"
- * leaves it out. Getting this wrong is a day of results, and the query
- * already says which is meant.
- *
- * @param {ParameterSlot} from
- * @param {ParameterSlot} to
- * @returns {{opening: string, joining: string}}
- */
-export function rangeWording(from, to) {
-  if (from.inclusive && to.inclusive) return { opening: 'Between', joining: 'and' };
-  if (from.inclusive) return { opening: 'Since', joining: 'and before' };
-  if (to.inclusive) return { opening: 'After', joining: 'until' };
-  return { opening: 'After', joining: 'and before' };
-}
-
-/**
- * The two slots of a range, lower end first, or null if the parameter is
- * not one — two equalities, say, which nothing can sensibly join.
- *
- * This asks {@link parameterRanges}, rather than deciding for itself, so
- * that the form and the check that follows it cannot disagree about what
- * a range is. Presenting two alternatives as "Between … and …" would read
- * as a period the query never meant, and one the check would then decline
- * to defend.
- *
- * @param {QueryParameter} parameter
- * @returns {[object, object]|null}
- */
-export function rangeEnds(parameter) {
-  const range = rangeOf(parameter);
-  return range ? [range.lower, range.upper] : null;
-}
-
-/**
- * The two ends of a parameter's range, lower first, or null if it is not
- * one.
- *
- * There are two ways to be a range, and the query says which in both.
- *
- * The author may declare it: two variables on one annotation line are the
- * start and the end, and nothing else needs deciding — they wrote the
- * query and know whether it asks for both.
- *
- * Or one variable may be bounded from each side by the same `&&`, which is
- * how SPARQL says the query needs both at once. Bounds that do not share a
- * conjunction are not paired, whatever else is true of them: separate
- * FILTERs, the two arms of a `||`, and opposite arms of a UNION all fail
- * that test without anything having to know what a UNION is.
- *
- * @param {QueryParameter} parameter
- * @returns {{lower: ParameterSlot, upper: ParameterSlot, strict: boolean}|null}
- *   where a strict range is one whose ends may not be equal, because at
- *   least one comparison excludes the value itself
- */
-function rangeOf(parameter) {
-  if (parameter.slots.length !== 2) return null;
-  const [first, second] = parameter.slots;
-
-  // Declared: the first variable is the start, whatever the query's own
-  // order. A declaration that does not hold makes no range at all, rather
-  // than some other pair of values wearing its label.
-  if (parameter.variables.length === 2) return parameter.declaredRange ? ends(first, second) : null;
-
-  if (first.conjunction === null || first.conjunction !== second.conjunction) return null;
-  const lower = parameter.slots.find(slot => slot.bound === 'lower');
-  const upper = parameter.slots.find(slot => slot.bound === 'upper');
-  return lower && upper ? ends(lower, upper) : null;
-}
-
-/** A pair of ends, with the wording their comparisons call for. */
-function ends(lower, upper) {
-  // `>` or `<` at either end leaves nothing between two equal values.
-  return { lower, upper, strict: !lower.inclusive || !upper.inclusive };
-}
-
-/**
- * Every range among a query's parameters: two values that would describe
- * an empty period if they were the wrong way round.
- *
- * @param {QueryParameter[]} parameters
- * @returns {Array<{lower: ParameterSlot, upper: ParameterSlot, strict: boolean}>}
- */
-export function parameterRanges(parameters) {
-  return parameters.map(rangeOf).filter(Boolean);
-}
-
-// ── values, in and out of the form ────────────────────────────────
-//
-// Which control can carry a value without altering it, what may be
-// typed into one, how two of them compare, and how one is written back
-// into the query.
-
-const DAY = /\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01])/.source;
-const CLOCK = /([01]\d|2[0-3]):[0-5]\d(:[0-5]\d)?/.source;
-
-const DISPLAYABLE = {
-  // Hour 24 is a valid xsd:time — the end of a day — and no clock control
-  // will take it. Nor will one take a month of 13, which is not valid
-  // either but can still be sitting in a query.
-  date: new RegExp(`^${DAY}$`),
-  time: new RegExp(`^${CLOCK}$`),
-  dateTime: new RegExp(`^${DAY}T${CLOCK}$`),
-  // A mantissa ending in a bare dot — `3.`, valid xsd:decimal — is
-  // cleared. Whether an exponent rescues it varies by browser, so any
-  // trailing dot is left to a text box, which always holds its value.
-  number: /^-?(\d+(\.\d+)?|\.\d+)([eE][+-]?\d+)?$/,
-  boolean: /^(true|false)$/,
-};
-
-/**
- * What SPARQL reads as a number written without quotes.
- *
- * Exactly the grammar's INTEGER, DECIMAL and DOUBLE, signed or not. It is
- * narrower than the lexical space of the datatypes: `3.` is a fine
- * xsd:decimal inside quotes, and not a number at all outside them, while
- * `INF` is only ever a quoted xsd:double.
- */
-const SPARQL_NUMBER =
-  /^[+-]?(\d+\.\d*[eE][+-]?\d+|\.\d+([eE][+-]?\d+)?|\d+(\.\d+)?([eE][+-]?\d+)?)$/;
-
-/**
- * What each kind that is written without quotes may be written as.
- *
- * A quoted value is safe whatever it holds, because escaping keeps it
- * inside its quotes. A bare one has no such wall: it becomes part of the
- * query itself, so it has to be a literal of that kind and nothing else.
- */
-const BARE = {
-  number: { shape: SPARQL_NUMBER, wanted: 'must be a number.' },
-  boolean: { shape: /^(true|false)$/, wanted: 'must be true or false.' },
-};
-
-/**
- * Why a value cannot be put in a slot, as a clause following the field's
- * name, or null if it can.
- *
- * @param {ParameterSlot} slot
- * @param {string} value
- * @returns {string|null}
- */
-export function valueProblem(slot, value) {
-  // Only a plain string may be empty, since `""` is a literal and
-  // `""^^xsd:date` is not a date.
-  if (value === '') return slot.kind === 'text' ? null : 'needs a value.';
-  // Quoting and escaping keep a value inside its literal whatever it says.
-  if (slot.quoted) return null;
-  const bare = BARE[slot.kind];
-  if (!bare) return 'cannot be written without quotes.';
-  return bare.shape.test(value) ? null : bare.wanted;
-}
-
-/**
- * The kind of control a slot should be given.
- *
- * Usually the kind of its value, but a value the control cannot hold is
- * offered as text instead. `"2024-11-04T10:00:00Z"` in a datetime-local
- * control would come back without its timezone, changing a query the user
- * never edited; a text box hands it back exactly as it was found.
- *
- * @param {ParameterSlot} slot
- * @returns {ParameterSlot['kind']}
- */
-export function controlKind(slot) {
-  const displayable = DISPLAYABLE[slot.kind];
-  if (displayable && !displayable.test(slot.value)) return 'text';
-  return slot.kind;
-}
-
-/**
- * A decimal broken into the pieces needed to compare it exactly: its sign,
- * its digits, and the power of ten they are scaled by.
- *
- * @param {string} value
- * @returns {{negative: boolean, digits: bigint, scale: number}|null}
- */
-function decimalParts(value) {
-  const match = value.match(/^([+-]?)(\d*)(?:\.(\d*))?(?:[eE]([+-]?\d+))?$/);
-  if (!match) return null;
-  const [, sign, whole = '', fraction = '', exponent] = match;
-  if (!whole && !fraction) return null;
-  const scale = Number(exponent || 0) - fraction.length;
-  // SPARQL puts no limit on how many digits an exponent may have, and past
-  // 308 of them a Number is Infinity. Two of those subtract to NaN, which
-  // no comparison against a limit rejects — every one is false — so the
-  // scale is refused here, where it is still a number.
-  if (!Number.isSafeInteger(scale)) return null;
-  return {
-    negative: sign === '-',
-    digits: BigInt(`${whole}${fraction}`),
-    scale,
-  };
-}
-
-/**
- * Which of two numbers is the smaller, exactly.
- *
- * Not through Number: XSD puts no limit on a decimal's precision, and past
- * 2^53 a Number stops telling one integer from the next — 9007199254740992
- * and 9007199254740993 become the same value, and an inverted range built
- * from them would pass unnoticed.
- *
- * @returns {-1|0|1|null} null where either value is not a plain number
- */
-function compareNumbers(a, b) {
-  const [x, y] = [decimalParts(a), decimalParts(b)];
-  if (!x || !y) return null;
-  if (x.digits === 0n && y.digits === 0n) return 0;
-  if (x.negative !== y.negative) return x.negative ? -1 : 1;
-
-  // Scale both to the same power of ten before comparing the digits. An
-  // exponent far beyond any real quantity is left unordered rather than
-  // turned into a number that long.
-  const difference = x.scale - y.scale;
-  if (Math.abs(difference) > 1000) return null;
-  const shift = 10n ** BigInt(Math.abs(difference));
-  const [left, right] = difference > 0 ? [x.digits * shift, y.digits] : [x.digits, y.digits * shift];
-
-  const order = left < right ? -1 : left > right ? 1 : 0;
-  return x.negative ? -order : order;
-}
-
-/**
- * Which of two values of the same kind comes first, or null when they
- * cannot be ordered soundly.
- *
- * Dates and times in their plain forms sort as text — that is what ISO
- * 8601 is for — but only once both carry the same precision, and not at
- * all once one carries a timezone. Numbers must be compared as numbers,
- * since "9" sorts after "10" and is not larger than it. Anything else,
- * including text, has no order worth asserting.
- *
- * @param {ParameterSlot['kind']} kind
- * @param {string} a
- * @param {string} b
- * @returns {-1|0|1|null}
- */
-export function compareValues(kind, a, b) {
-  if (kind === 'number') return compareNumbers(a, b);
-
-  const shape = kind === 'date' || kind === 'time' || kind === 'dateTime' ? DISPLAYABLE[kind] : null;
-  if (!shape || !shape.test(a) || !shape.test(b)) return null;
-
-  // A control may hand back a time without seconds; the same instant must
-  // not sort before itself written more precisely.
-  const seconds = (value) => (kind === 'date' || /:\d{2}:\d{2}/.test(value) ? value : `${value}:00`);
-  const [x, y] = [seconds(a), seconds(b)];
-  return x < y ? -1 : x > y ? 1 : 0;
-}
-
-/**
- * A slot's literal, written with a new value.
- *
- * @param {ParameterSlot} slot
- * @param {string} value
- * @returns {string}
- */
-export function literalFor(slot, value) {
-  const lexical = lexicalForm(slot, String(value));
-  // A number or a boolean is written bare; quoting it would change its type.
-  if (slot.quoted) return `"${escapeLiteral(lexical)}"${slot.suffix}`;
-
-  // Bare, so the value becomes part of the query rather than sitting inside
-  // it. `1) . ?x ?p ?o . FILTER(1` is a query of its own, and it parses.
-  // The form refuses such a value long before this, and this refuses it
-  // again: a wrong query is worth an error, never a run.
-  if (!BARE[slot.kind]?.shape.test(lexical)) {
-    throw new Error(`${JSON.stringify(lexical)} is not a ${slot.kind} SPARQL can read unquoted`);
-  }
-  return lexical;
-}
-
-/**
- * A value in the form its datatype requires.
- *
- * A `datetime-local` control yields `2024-11-04T00:00`, and xsd:dateTime
- * has no such form — seconds are part of it. Writing the control's value
- * straight into the query would replace a valid literal with an invalid
- * one, which is the failure this whole area exists to prevent.
- *
- * @param {ParameterSlot} slot
- * @param {string} value
- * @returns {string}
- */
-function lexicalForm(slot, value) {
-  if (slot.kind === 'dateTime' && /^-?\d{4,}-\d{2}-\d{2}T\d{2}:\d{2}$/.test(value)) {
-    return `${value}:00`;
-  }
-  if (slot.kind === 'time' && /^\d{2}:\d{2}$/.test(value)) return `${value}:00`;
-  return value;
-}
-
-/**
- * The query with the given values in place of its own.
- *
- * @param {string} text the query as it stands
- * @param {QueryParameter[]} parameters as {@link queryParameters} found them
- * @param {Map<ParameterSlot, string>} values the value for each slot; a slot
- *   with no entry keeps the value already in the query
- * @returns {string}
- */
-export function fillQuery(text, parameters, values) {
-  const edits = [];
-  for (const parameter of parameters) {
-    for (const slot of parameter.slots) {
-      if (!values.has(slot)) continue;
-      edits.push({ from: slot.from, to: slot.to, insert: literalFor(slot, values.get(slot)) });
-    }
-  }
-  return applyEdits(text, edits);
 }
